chore: refactor for re-use
