+++ frontend/src/app/components/containers/parties/edit/party-edit.tsx
@@ -22,19 +22,16 @@ import {
   PersonInput,
   PersonUpdateInput,
 } from "@/generated/graphql";
-import { CompInput } from "@/app/components/common/comp-input";
 import { selectPartyTypeDropdown } from "@/app/store/reducers/code-table-selectors";
-import { Button } from "react-bootstrap";
 import { GET_PARTY } from "@/app/components/containers/parties/view/party-view";
-import { ContactPersonFields } from "@/app/components/containers/parties/edit/contact-person";
-import { ValidationPhoneInput } from "@/app/common/validation-phone-input";
-import { ValidationDatePicker } from "@/app/common/validation-date-picker";
 import { selectSexDropdown } from "@/app/store/reducers/code-table";
 import { parse } from "date-fns";
 import useUnsavedChangesWarning from "@/app/hooks/use-unsaved-changes-warning";
 import { ContactMethods } from "@/app/constants/contact-methods";
 import { BusinessIdentifiers } from "@/app/constants/business-identifiers";
 import { PartyTypeCodes } from "@/app/constants/party-types";
+import { PersonForm } from "@/app/components/containers/parties/form/person-form";
+import { BusinessFormFields } from "@/app/components/containers/parties/form/business-form";
 
 const PARTY_PERSON_FRAGMENT = gql`
   fragment PartyPersonFields on Person {
@@ -319,10 +316,6 @@ const PartyEdit: FC = () => {
       };
     });
 
-  const sexCodeDropdownOptions = sexCodeOptions
-    ?.filter((opt: { isActive?: boolean }) => opt.isActive !== false)
-    .map((opt: { value: string; label: string }) => ({ value: opt.value, label: opt.label }));
-
   const createPartyMutation = useGraphQLMutation(CREATE_PARTY_MUTATION, {
     onError: (error: any) => {
       console.error("Error creating party:", error);
@@ -428,22 +421,11 @@ const PartyEdit: FC = () => {
   const { allowNavigation } = useUnsavedChangesWarning(isDirty);
 
   const partyTypeValue = useStore(form.store, (state) => state.values.partyType);
-  const aliasesValue = useStore(form.store, (state) => state.values.aliases);
-  const phoneNumberValue = useStore(form.store, (state) => state.values.phoneNumbers);
-  const emailAddressValue = useStore(form.store, (state) => state.values.emailAddresses);
-  const contactValue = useStore(form.store, (state) => state.values.contacts);
 
   const navigateToPartyList = () => {
     allowNavigation();
     navigate(`/parties`);
   };
-
-  const focusFieldById = useCallback((fieldId: string) => {
-    setTimeout(() => {
-      const field = document.getElementById(fieldId);
-      field?.focus();
-    }, 0);
-  }, []);
 
   const confirmCancelChanges = useCallback(() => {
     form.reset();
@@ -473,329 +455,6 @@ const PartyEdit: FC = () => {
   const saveButtonClick = useCallback(() => {
     form.handleSubmit();
   }, [form]);
-
-  const handleAddAlias = useCallback(() => {
-    const currentAliases = form.getFieldValue("aliases") || [];
-    const newAliases = [...currentAliases, { aliasGuid: crypto.randomUUID(), name: "" }];
-    form.setFieldValue("aliases", newAliases);
-    focusFieldById(`alias-${currentAliases.length}`);
-  }, [form]);
-
-  const handleRemoveAlias = useCallback(
-    (indexToRemove: number) => {
-      const currentAliases = form.getFieldValue("aliases") || [];
-      const newAliases = currentAliases.filter((_: Alias, index: number) => index !== indexToRemove);
-      form.setFieldValue("aliases", newAliases);
-    },
-    [form],
-  );
-
-  const handleAddPhoneNumber = useCallback(() => {
-    const currentPhoneNumbers = form.getFieldValue("phoneNumbers") || [];
-    const newPhoneNumbers = [
-      ...currentPhoneNumbers,
-      {
-        contactMethodGuid: "",
-        value: "",
-        isPrimary: currentPhoneNumbers.length === 0, // First one is primary
-      },
-    ];
-    form.setFieldValue("phoneNumbers", newPhoneNumbers);
-    focusFieldById(`phone-number-${currentPhoneNumbers.length}`);
-  }, [form]);
-
-  const handleRemovePhoneNumber = useCallback(
-    (indexToRemove: number) => {
-      const currentPhoneNumbers = form.getFieldValue("phoneNumbers") || [];
-      const removingPrimary = currentPhoneNumbers[indexToRemove]?.isPrimary;
-      const newPhoneNumbers = currentPhoneNumbers.filter((_: any, index: number) => index !== indexToRemove);
-
-      // If we removed the primary phone and there are still phone numbers left,
-      // make the first one primary
-      if (removingPrimary && newPhoneNumbers.length > 0) {
-        newPhoneNumbers[0].isPrimary = true;
-      }
-
-      form.setFieldValue("phoneNumbers", newPhoneNumbers);
-    },
-    [form],
-  );
-
-  const handleAddEmail = useCallback(() => {
-    const currentEmails = form.getFieldValue("emailAddresses") || [];
-    const newEmails = [
-      ...currentEmails,
-      {
-        contactMethodGuid: "",
-        value: "",
-        isPrimary: currentEmails.length === 0, // First one is primary
-      },
-    ];
-    form.setFieldValue("emailAddresses", newEmails);
-    focusFieldById(`email-${currentEmails.length}`);
-  }, [form]);
-
-  const handleRemoveEmail = useCallback(
-    (indexToRemove: number) => {
-      const currentEmails = form.getFieldValue("emailAddresses") || [];
-      const removingPrimary = currentEmails[indexToRemove]?.isPrimary;
-      const newEmails = currentEmails.filter((_: any, index: number) => index !== indexToRemove);
-
-      // If we removed the primary phone and there are still phone numbers left,
-      // make the first one primary
-      if (removingPrimary && newEmails.length > 0) {
-        newEmails[0].isPrimary = true;
-      }
-
-      form.setFieldValue("emailAddresses", newEmails);
-    },
-    [form],
-  );
-
-  const handleAddContact = useCallback(() => {
-    const currentContacts = form.getFieldValue("contacts") || [];
-    const newContact = {
-      businessPersonXrefGuid: undefined,
-      business: {
-        businessGuid: partyData?.party?.business?.businessGuid || "",
-      },
-      person: {
-        personGuid: "",
-        firstName: "",
-        lastName: "",
-        contactMethods: [],
-      },
-    };
-    form.setFieldValue("contacts", [...currentContacts, newContact]);
-    focusFieldById(`contact-firstName-${currentContacts.length}`);
-  }, [form]);
-
-  const handleRemoveContact = useCallback(
-    (indexToRemove: number) => {
-      const currentContacts = form.getFieldValue("contacts") || [];
-      const newContacts = currentContacts.filter((_: BusinessPerson, index: number) => index !== indexToRemove);
-      form.setFieldValue("contacts", newContacts);
-    },
-    [form],
-  );
-
-  const handleAddContactMethod = useCallback(
-    (contactIndex: number, typeCode: string) => {
-      const currentContacts = form.getFieldValue("contacts") || [];
-
-      // Check if this will be the first contact method of this type for this contact
-      const existingMethodsOfType =
-        currentContacts[contactIndex]?.person?.contactMethods?.filter(
-          (cm: ContactMethod) => cm?.typeCode === typeCode,
-        ) || [];
-
-      const newContactMethod = {
-        contactMethodGuid: undefined, // ← Should be undefined, not empty string
-        typeCode: typeCode,
-        value: "",
-        isPrimary: existingMethodsOfType.length === 0, // First one is primary
-      };
-
-      const updatedContacts = currentContacts.map((contact: BusinessPerson, index: number) => {
-        if (index === contactIndex) {
-          return {
-            businessPersonXrefGuid: contact.businessPersonXrefGuid,
-            business: { businessGuid: contact.business?.businessGuid },
-            person: {
-              personGuid: contact.person?.personGuid,
-              firstName: contact.person?.firstName,
-              lastName: contact.person?.lastName,
-              contactMethods: [
-                ...(contact.person?.contactMethods || [])
-                  .filter((cm): cm is ContactMethod => cm != null)
-                  .map((cm) => ({
-                    contactMethodGuid: cm.contactMethodGuid,
-                    typeCode: cm.typeCode,
-                    value: cm.value,
-                    isPrimary: cm.isPrimary ?? false,
-                  })),
-                newContactMethod,
-              ],
-            },
-          };
-        }
-        return {
-          businessPersonXrefGuid: contact.businessPersonXrefGuid,
-          business: { businessGuid: contact.business?.businessGuid },
-          person: {
-            personGuid: contact.person?.personGuid,
-            firstName: contact.person?.firstName,
-            lastName: contact.person?.lastName,
-            contactMethods: (contact.person?.contactMethods || [])
-              .filter((cm): cm is ContactMethod => cm != null)
-              .map((cm) => ({
-                contactMethodGuid: cm.contactMethodGuid,
-                typeCode: cm.typeCode,
-                value: cm.value,
-                isPrimary: cm.isPrimary ?? false,
-              })),
-          },
-        };
-      });
-
-      form.setFieldValue("contacts", updatedContacts);
-
-      const fieldId =
-        typeCode === ContactMethods.PHONE
-          ? `contact-phone-${contactIndex}-${existingMethodsOfType.length}`
-          : `contact-email-${contactIndex}-${existingMethodsOfType.length}`;
-      focusFieldById(fieldId);
-    },
-    [form],
-  );
-
-  const handleRemoveContactMethod = useCallback(
-    (contactIndex: number, methodIndex: number) => {
-      const currentContacts = form.getFieldValue("contacts") || [];
-
-      const updatedContacts = currentContacts.map((contact: BusinessPerson, index: number) => {
-        if (index === contactIndex) {
-          const contactMethods = (contact.person?.contactMethods || []).filter((cm): cm is ContactMethod => cm != null);
-
-          const methodToRemove = contactMethods[methodIndex];
-          const removingPrimary = methodToRemove?.isPrimary || false;
-          const removingTypeCode = methodToRemove?.typeCode;
-          let newContactMethods = contactMethods
-            .filter((_, i) => i !== methodIndex)
-            .map((cm) => ({
-              contactMethodGuid: cm.contactMethodGuid,
-              typeCode: cm.typeCode,
-              value: cm.value,
-              isPrimary: cm.isPrimary ?? false,
-            }));
-
-          // If we removed the primary method and there are still methods of the same type left,
-          // make the first one of that type primary
-          if (removingPrimary && removingTypeCode && newContactMethods.length > 0) {
-            const firstOfSameType = newContactMethods.findIndex((cm) => cm?.typeCode === removingTypeCode);
-            if (firstOfSameType !== -1) {
-              newContactMethods[firstOfSameType] = {
-                ...newContactMethods[firstOfSameType],
-                isPrimary: true,
-              };
-            }
-          }
-
-          return {
-            businessPersonXrefGuid: contact.businessPersonXrefGuid,
-            business: { businessGuid: contact.business?.businessGuid },
-            person: {
-              personGuid: contact.person?.personGuid,
-              firstName: contact.person?.firstName,
-              lastName: contact.person?.lastName,
-              contactMethods: newContactMethods,
-            },
-          };
-        }
-        return {
-          businessPersonXrefGuid: contact.businessPersonXrefGuid,
-          business: { businessGuid: contact.business?.businessGuid },
-          person: {
-            personGuid: contact.person?.personGuid,
-            firstName: contact.person?.firstName,
-            lastName: contact.person?.lastName,
-            contactMethods: (contact.person?.contactMethods || [])
-              .filter((cm): cm is ContactMethod => cm != null)
-              .map((cm) => ({
-                contactMethodGuid: cm.contactMethodGuid,
-                typeCode: cm.typeCode,
-                value: cm.value,
-                isPrimary: cm.isPrimary ?? false,
-              })),
-          },
-        };
-      });
-
-      form.setFieldValue("contacts", updatedContacts);
-    },
-    [form],
-  );
-
-  const handleSetPrimaryPhoneNumber = useCallback(
-    (index: number) => {
-      const currentPhoneNumbers = form.getFieldValue("phoneNumbers") || [];
-      const updatedPhones = currentPhoneNumbers.map((p: any, i: number) => ({
-        ...p,
-        isPrimary: i === index,
-      }));
-      form.setFieldValue("phoneNumbers", updatedPhones);
-    },
-    [form],
-  );
-
-  const handleSetPrimaryEmail = useCallback(
-    (index: number) => {
-      const currentEmailAddresses = form.getFieldValue("emailAddresses") || [];
-      const updatedEmailAddresses = currentEmailAddresses.map((p: any, i: number) => ({
-        ...p,
-        isPrimary: i === index,
-      }));
-      form.setFieldValue("emailAddresses", updatedEmailAddresses);
-    },
-    [form],
-  );
-
-  const handleSetPrimaryContact = useCallback(
-    (contactIndex: number, contactMethodIndex: number, contactMethodType: string) => {
-      const currentContacts = form.getFieldValue("contacts") || [];
-      const updatedContacts = currentContacts.map((c: BusinessPerson, cIndex: number) => {
-        if (cIndex === contactIndex) {
-          return {
-            businessPersonXrefGuid: c.businessPersonXrefGuid,
-            business: { businessGuid: c.business?.businessGuid },
-            person: {
-              personGuid: c.person?.personGuid,
-              firstName: c.person?.firstName,
-              lastName: c.person?.lastName,
-              contactMethods: (c.person?.contactMethods || [])
-                .filter((cm): cm is ContactMethod => cm != null)
-                .map((cm, cmIdx) => {
-                  if (cm?.typeCode === contactMethodType) {
-                    return {
-                      contactMethodGuid: cm.contactMethodGuid,
-                      typeCode: cm.typeCode,
-                      value: cm.value,
-                      isPrimary: cmIdx === contactMethodIndex,
-                    };
-                  }
-                  return {
-                    contactMethodGuid: cm.contactMethodGuid,
-                    typeCode: cm.typeCode,
-                    value: cm.value,
-                    isPrimary: cm.isPrimary ?? false, // ← Add default
-                  };
-                }),
-            },
-          };
-        }
-        // Transform the unchanged contact too
-        return {
-          businessPersonXrefGuid: c.businessPersonXrefGuid,
-          business: { businessGuid: c.business?.businessGuid },
-          person: {
-            personGuid: c.person?.personGuid,
-            firstName: c.person?.firstName,
-            lastName: c.person?.lastName,
-            contactMethods: (c.person?.contactMethods || [])
-              .filter((cm): cm is ContactMethod => cm != null)
-              .map((cm) => ({
-                contactMethodGuid: cm.contactMethodGuid,
-                typeCode: cm.typeCode,
-                value: cm.value,
-                isPrimary: cm.isPrimary ?? false, // ← Add default
-              })),
-          },
-        };
-      });
-      form.setFieldValue("contacts", updatedContacts);
-    },
-    [form],
-  );
 
   const isSubmitting = createPartyMutation.isPending || updatePartyMutation.isPending;
   const isDisabled = isSubmitting || isLoading;
@@ -840,507 +499,18 @@ const PartyEdit: FC = () => {
               )}
             />
             {partyTypeValue === "PRS" && (
-              <>
-                <FormField
-                  form={form}
-                  name="firstName"
-                  label="First name"
-                  required
-                  validators={{ onChange: z.string().min(1, "First name is required") }}
-                  render={(field) => (
-                    <CompInput
-                      id="FirstName"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      defaultValue={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={50}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      placeholder="Enter first name..."
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="lastName"
-                  label="Last name"
-                  required
-                  validators={{ onChange: z.string().min(1, "Last name is required") }}
-                  render={(field) => (
-                    <CompInput
-                      id="LastName"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      defaultValue={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={50}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      placeholder="Enter last name..."
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="middleName"
-                  label="Middle name"
-                  render={(field) => (
-                    <CompInput
-                      id="MiddleName"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      defaultValue={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={50}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      placeholder="Enter middle name..."
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="middleName2"
-                  label="Middle name 2"
-                  render={(field) => (
-                    <CompInput
-                      id="MiddleName2"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      defaultValue={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={50}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      placeholder="Enter middle name 2..."
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="dateOfBirth"
-                  label="Date of birth"
-                  render={(field) => (
-                    <ValidationDatePicker
-                      id="DateOfBirth"
-                      classNamePrefix="comp-details-input"
-                      className="comp-form-control comp-details-input"
-                      selectedDate={field.state.value}
-                      onChange={(date: Date | undefined) => field.handleChange(date ?? undefined)}
-                      errMsg={field.state.meta.errors?.[0]?.message || ""}
-                      isDisabled={isDisabled}
-                      showYearDropdown={true}
-                      yearDropdownItemNumber={100}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="driversLicenseNumber"
-                  label="Driver's licence number"
-                  render={(field) => (
-                    <CompInput
-                      id="DriversLicenseNumber"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      defaultValue={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={50}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      placeholder="Enter driver's licence number..."
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="driversLicenseJurisdiction"
-                  label="Driver's licence jurisdiction"
-                  render={(field) => (
-                    <CompInput
-                      id="DriversLicenseJurisdiction"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      defaultValue={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={50}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      placeholder="Enter driver's licence jurisdiction..."
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="sexCode"
-                  label="Sex"
-                  render={(field) => (
-                    <CompSelect
-                      id="sex-select"
-                      classNamePrefix="comp-select"
-                      className="comp-details-input"
-                      options={sexCodeDropdownOptions}
-                      value={sexCodeDropdownOptions?.find((opt: any) => opt.value === field.state.value)}
-                      onChange={(option) => field.handleChange(option?.value ?? "")}
-                      placeholder="Select sex"
-                      isClearable={true}
-                      showInactive={false}
-                      enableValidation={true}
-                      errorMessage={field.state.meta.errors?.[0]?.message || ""}
-                      isDisabled={isDisabled}
-                    />
-                  )}
-                />
-                {phoneNumberValue?.map((phoneNumber: ContactMethod, index: number) => (
-                  <FormField
-                    key={phoneNumber.contactMethodGuid || `phone-${index}`}
-                    form={form}
-                    name={`phoneNumbers[${index}].value`}
-                    label={index === 0 ? "Phone number" : ""}
-                    render={(field) => (
-                      <div className="party-contact-method">
-                        {index === 0 && <div className="party-primary-contact-method-label">Primary</div>}
-                        {index > 0 && <div className="party-primary-contact-spacer"></div>}
-
-                        <input
-                          type="radio"
-                          id={`person-phone-primary-${index}`}
-                          name="primaryPhoneNumber"
-                          checked={phoneNumber.isPrimary || false}
-                          onChange={() => handleSetPrimaryPhoneNumber(index)}
-                          disabled={isDisabled}
-                        />
-
-                        <div className="party-multiple-value-container">
-                          <ValidationPhoneInput
-                            className="comp-details-input"
-                            value={phoneNumber.value ?? ""}
-                            onChange={(value: string) => field.handleChange(value || "")}
-                            maxLength={14}
-                            international={false}
-                            id={`person-phone-number-${index}`}
-                            errMsg={field.state.meta.errors?.[0]?.message || ""}
-                          />
-                        </div>
-
-                        <Button
-                          variant="outline-primary"
-                          size="sm"
-                          onClick={() => handleRemovePhoneNumber(index)}
-                          type="button"
-                        >
-                          <i className="bi bi-trash" />
-                          {/**/}
-                          Remove
-                        </Button>
-                      </div>
-                    )}
-                  />
-                ))}
-                <FormField
-                  form={form}
-                  name="add-person-phone-number-placeholder"
-                  label=""
-                  render={() => (
-                    <Button
-                      id="add-person-phone-number-button"
-                      variant="outline-primary"
-                      size="sm"
-                      onClick={handleAddPhoneNumber}
-                      type="button"
-                    >
-                      <i className="bi bi-plus-circle me-1" /> {/**/}
-                      Add phone number
-                    </Button>
-                  )}
-                />
-              </>
+              <PersonForm
+                form={form}
+                isDisabled={isDisabled}
+              />
             )}
             {partyTypeValue === "CMP" && (
-              <>
-                <FormField
-                  form={form}
-                  name="businessName"
-                  label="Name"
-                  required
-                  validators={{
-                    onChange: z.string().min(1, "Name is required"),
-                  }}
-                  render={(field) => (
-                    <CompInput
-                      id="businessName"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      value={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={50}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      placeholder="Enter name..."
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                {aliasesValue?.map((alias: Alias, index: number) => (
-                  <FormField
-                    key={alias.aliasGuid || `alias-${index}`}
-                    form={form}
-                    name={`aliases[${index}].name` as any}
-                    label={index === 0 ? "Alias" : ""}
-                    render={(field) => (
-                      <div className="party-alias-container">
-                        <div className="party-multiple-value-container">
-                          <CompInput
-                            id={`alias-${index}`}
-                            divid=""
-                            type="input"
-                            inputClass="comp-form-control comp-details-input"
-                            value={field.state.value}
-                            error={field.state.meta.errors?.[0]?.message || ""}
-                            maxLength={512}
-                            onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                            placeholder="Enter alias..."
-                            disabled={isDisabled}
-                          />
-                        </div>
-
-                        <Button
-                          variant="outline-primary"
-                          size="sm"
-                          onClick={() => handleRemoveAlias(index)}
-                          type="button"
-                        >
-                          <i className="bi bi-trash" /> {/**/}
-                          Remove
-                        </Button>
-                      </div>
-                    )}
-                  />
-                ))}
-                <FormField
-                  form={form}
-                  name="add-alias-placeholder"
-                  label=""
-                  render={() => (
-                    <Button
-                      id="add-task-button"
-                      variant="outline-primary"
-                      size="sm"
-                      onClick={handleAddAlias}
-                      type="button"
-                    >
-                      <i className="bi bi-plus-circle me-1" /> {/**/}
-                      Add alias
-                    </Button>
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="businessNumber.identifierValue"
-                  label="Business number"
-                  render={(field) => (
-                    <CompInput
-                      id="businessNumber"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      value={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={16}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                <FormField
-                  form={form}
-                  name="worksafeBCNumber.identifierValue"
-                  label="WorkSafeBC number"
-                  render={(field) => (
-                    <CompInput
-                      id="worksafeBCNumber"
-                      divid=""
-                      type="input"
-                      inputClass="comp-form-control comp-details-input"
-                      value={field.state.value}
-                      error={field.state.meta.errors?.[0]?.message || ""}
-                      maxLength={16}
-                      onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                      disabled={isDisabled}
-                    />
-                  )}
-                />
-                {phoneNumberValue?.map((phoneNumber: ContactMethod, index: number) => (
-                  <FormField
-                    key={phoneNumber.contactMethodGuid || `phone-${index}`}
-                    form={form}
-                    name={`phoneNumbers[${index}].value`}
-                    label={index === 0 ? "Phone number" : ""}
-                    render={(field) => (
-                      <div className="party-contact-method">
-                        {index === 0 && <div className="party-primary-contact-method-label">Primary</div>}
-                        {index > 0 && <div className="party-primary-contact-spacer"></div>}
-
-                        <input
-                          type="radio"
-                          id={`phone-primary-${index}`}
-                          name="primaryPhoneNumber"
-                          checked={phoneNumber.isPrimary || false}
-                          onChange={() => handleSetPrimaryPhoneNumber(index)}
-                          disabled={isDisabled}
-                        />
-
-                        <div className="party-multiple-value-container">
-                          <ValidationPhoneInput
-                            className="comp-details-input"
-                            value={phoneNumber.value ?? ""}
-                            onChange={(value: string) => field.handleChange(value || "")}
-                            maxLength={14}
-                            international={false}
-                            id={`phone-number-${index}`}
-                            errMsg={field.state.meta.errors?.[0]?.message || ""}
-                          />
-                        </div>
-
-                        <Button
-                          variant="outline-primary"
-                          size="sm"
-                          onClick={() => handleRemovePhoneNumber(index)}
-                          type="button"
-                        >
-                          <i className="bi bi-trash" />
-                          {/**/}
-                          Remove
-                        </Button>
-                      </div>
-                    )}
-                  />
-                ))}
-                <FormField
-                  form={form}
-                  name="add-phone-number-placeholder"
-                  label=""
-                  render={() => (
-                    <Button
-                      id="add-phone-number-button"
-                      variant="outline-primary"
-                      size="sm"
-                      onClick={handleAddPhoneNumber}
-                      type="button"
-                    >
-                      <i className="bi bi-plus-circle me-1" /> {/**/}
-                      Add phone number
-                    </Button>
-                  )}
-                />
-                {emailAddressValue?.map((email: ContactMethod, index: number) => (
-                  <FormField
-                    key={email.contactMethodGuid || `email-${index}`}
-                    form={form}
-                    name={`emailAddresses[${index}].value` as any}
-                    label={index === 0 ? "Email" : ""}
-                    render={(field) => (
-                      <div className="party-contact-method">
-                        {index === 0 && <div className="party-primary-contact-method-label">Primary</div>}
-                        {index > 0 && <div className="party-primary-contact-spacer"></div>}
-
-                        <input
-                          type="radio"
-                          id={`email-primary-${index}`}
-                          name="primaryEmail"
-                          checked={email.isPrimary || false}
-                          onChange={() => handleSetPrimaryEmail(index)}
-                          disabled={isDisabled}
-                        />
-
-                        <div className="party-multiple-value-container">
-                          <CompInput
-                            id={`email-${index}`}
-                            divid=""
-                            type="input"
-                            inputClass="comp-form-control comp-details-input"
-                            value={field.state.value}
-                            error={field.state.meta.errors?.[0]?.message || ""}
-                            maxLength={512}
-                            onChange={(evt: any) => field.handleChange(evt?.target?.value || "")}
-                            disabled={isDisabled}
-                          />
-                        </div>
-                        <Button
-                          variant="outline-primary"
-                          size="sm"
-                          onClick={() => handleRemoveEmail(index)}
-                          type="button"
-                        >
-                          <i className="bi bi-trash" />
-                          {/**/}
-                          Remove
-                        </Button>
-                      </div>
-                    )}
-                  />
-                ))}
-                <FormField
-                  form={form}
-                  name="add-email-placeholder"
-                  label=""
-                  render={() => (
-                    <Button
-                      id="add-email-button"
-                      variant="outline-primary"
-                      size="sm"
-                      onClick={handleAddEmail}
-                      type="button"
-                    >
-                      <i className="bi bi-plus-circle me-1" /> {/**/}
-                      Add email
-                    </Button>
-                  )}
-                />
-                {contactValue?.map((contact: BusinessPerson, contactIndex: number) => (
-                  <FormField
-                    key={contact.businessPersonXrefGuid || `contact-${contactIndex}`}
-                    form={form}
-                    name={`contact-${contactIndex}.person`}
-                    label={contactIndex === 0 ? "Contact" : ""}
-                    render={() => (
-                      <ContactPersonFields
-                        contact={contact}
-                        contactIndex={contactIndex}
-                        form={form}
-                        isDisabled={isDisabled}
-                        onRemoveContact={handleRemoveContact}
-                        onAddContactMethod={handleAddContactMethod}
-                        onRemoveContactMethod={handleRemoveContactMethod}
-                        onSetPrimaryContact={handleSetPrimaryContact}
-                      />
-                    )}
-                  />
-                ))}
-
-                <FormField
-                  form={form}
-                  name="add-email-placeholder"
-                  label=""
-                  render={() => (
-                    <Button
-                      variant="outline-primary"
-                      onClick={handleAddContact}
-                      size="sm"
-                      type="button"
-                    >
-                      <i className="bi bi-plus-circle me-1" /> Add contact
-                    </Button>
-                  )}
-                />
-              </>
+              <BusinessFormFields
+                form={form}
+                isDisabled={isDisabled}
+                showContactPeople={true}
+                businessGuid={partyData?.party?.business?.businessGuid}
+              />
             )}
           </fieldset>
         </form>
